Code

Code

Functions

Tokamak.Area_Cable — Function.

Area_Cable(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Area of total cable.

source

Tokamak.Area_Cu — Function.

Area_Cu(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Determine Copper Stablizer Dimensions.

source

Tokamak.Area_H2 — Function.

Area_H2(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Determine Cooling Channel Dimensions.

source

Tokamak.Area_HTS — Method.

Area_HTS()

Area of HTS in winding pack.

source

Tokamak.Area_St — Function.

Area_St(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Determine Steel Structure Dimensions.

source

Tokamak.Area_Tape — Method.

Area_Tape()

Lorem ipsum dolor sit amet.

source

Tokamak.B0pnew — Method.

B0pnew()

Lorem ipsum dolor sit amet.

source

Tokamak.B_M — Method.

B_M(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.B_coil_max — Method.

B_coil_max()

Lorem ipsum dolor sit amet.

source

Tokamak.Bvtotnew_SF — Method.

Bvtotnew_SF()

Lorem ipsum dolor sit amet.

source

Tokamak.BzDivs — Method.

BzDivs()

Divertor Field Contribution.

source

Tokamak.BzDivsnew — Method.

BzDivsnew()

Calculate New BV due to Divertor.

source

Tokamak.BzPFnew — Method.

BzPFnew()

Lorem ipsum dolor sit amet.

source

Tokamak.BzPFs — Method.

BzPFs()

Lorem ipsum dolor sit amet.

source

Tokamak.BzPFsnew — Method.

BzPFsnew()

Lorem ipsum dolor sit amet.

source

Tokamak.Bzfunc — Method.

Bzfunc(rho, eta, curB, k)

Bz Field Contribution.

source

Tokamak.C_B — Method.

C_B()

Lorem ipsum dolor sit amet.

source

Tokamak.C_TF — Method.

C_TF()

Lorem ipsum dolor sit amet.

source

Tokamak.Cable_L — Function.

Cable_L(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Length of one cable [m].

source

Tokamak.Coil_Energy — Function.

Coil_Energy(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

TF coil magnetic energy.

source

Tokamak.Coil_Inductance — Function.

Coil_Inductance(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

TF coil inductance.

source

Tokamak.Coil_P — Function.

Coil_P(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Coil perimeter [m].

source

Tokamak.Cost_CS — Function.

Cost_CS(cur_solution=solve_magnet_equations())

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_Cu — Method.

Cost_Cu()

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_HTS_Total — Function.

Cost_HTS_Total(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=nothing)

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_PF — Method.

Cost_PF()

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_ST_Total — Function.

Cost_ST_Total(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=nothing)

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_St — Method.

Cost_St()

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_TF — Method.

Cost_TF()

Lorem ipsum dolor sit amet.

source

Tokamak.Cost_Tape — Method.

Cost_Tape()

Lorem ipsum dolor sit amet.

source

Tokamak.Cryo_Operation — Function.

Cryo_Operation(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=nothing)

Cost of power @ 0.1 /kWhr to power the cryoplant annually.

source

Tokamak.Cryoplant_Cost — Function.

Cryoplant_Cost(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=nothing)

Cost of cryoplant scaled to various heat loads, scaled to 4.5K and estimated using Green's Formula.

source

Tokamak.Cryostat_Cost — Method.

Cryostat_Cost()

Cost of cryostat scaled to ITER.

source

Tokamak.Cu_Cp — Method.

Cu_Cp(T)

Specific heat of copper as func of temperature.

source

Tokamak.E_F — Method.

E_F()

Lorem ipsum dolor sit amet.

source

Tokamak.E_n — Method.

E_n()

Lorem ipsum dolor sit amet.

source

Tokamak.Frac_Cu — Method.

Frac_Cu()

Fraction of copper.

source

Tokamak.Frac_H2 — Method.

Frac_H2()

Fraction of H2.

source

Tokamak.Frac_HTS — Function.

Frac_HTS(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Fraction of HTS.

source

Tokamak.Frac_St — Method.

Frac_St()

Fraction of steel.

source

Tokamak.I_CD — Method.

I_CD()

Lorem ipsum dolor sit amet.

source

Tokamak.ItPF — Method.

ItPF()

Lorem ipsum dolor sit amet.

source

Tokamak.K_B — Method.

K_B()

Lorem ipsum dolor sit amet.

source

Tokamak.K_BR — Method.

K_BR()

Lorem ipsum dolor sit amet.

source

Tokamak.K_CD — Method.

K_CD()

Lorem ipsum dolor sit amet.

source

Tokamak.K_CD_hat — Method.

K_CD_hat()

Lorem ipsum dolor sit amet.

source

Tokamak.K_DV — Method.

K_DV()

Lorem ipsum dolor sit amet.

source

Tokamak.K_F — Method.

K_F()

Lorem ipsum dolor sit amet.

source

Tokamak.K_H — Method.

K_H()

Lorem ipsum dolor sit amet.

source

Tokamak.K_I — Method.

K_I()

Lorem ipsum dolor sit amet.

source

Tokamak.K_L — Method.

K_L()

Lorem ipsum dolor sit amet.

source

Tokamak.K_N_G — Method.

K_N_G()

Lorem ipsum dolor sit amet.

source

Tokamak.K_PB — Method.

K_PB()

Lorem ipsum dolor sit amet.

source

Tokamak.K_W — Method.

K_W()

Lorem ipsum dolor sit amet.

source

Tokamak.K_beta — Method.

K_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.K_kappa — Method.

K_kappa()

Lorem ipsum dolor sit amet.

source

Tokamak.K_kernel — Method.

K_kernel()

Lorem ipsum dolor sit amet.

source

Tokamak.K_n — Method.

K_n()

Lorem ipsum dolor sit amet.

source

Tokamak.K_rad — Method.

K_rad()

Lorem ipsum dolor sit amet.

source

Tokamak.MWh_electric_per_year — Method.

MWh_electric_per_year()

Discount electricity per year.

source

Tokamak.P_BR — Method.

P_BR()

Lorem ipsum dolor sit amet.

source

Tokamak.P_F — Method.

P_F()

Lorem ipsum dolor sit amet.

source

Tokamak.P_H — Method.

P_H()

Lorem ipsum dolor sit amet.

source

Tokamak.P_T — Method.

P_T()

Lorem ipsum dolor sit amet.

source

Tokamak.P_alpha — Method.

P_alpha()

Lorem ipsum dolor sit amet.

source

Tokamak.P_kappa — Method.

P_kappa()

Lorem ipsum dolor sit amet.

source

Tokamak.P_n — Method.

P_n()

Lorem ipsum dolor sit amet.

source

Tokamak.Q_E — Method.

Q_E()

Lorem ipsum dolor sit amet.

source

Tokamak.S_W — Method.

S_W()

Lorem ipsum dolor sit amet.

source

Tokamak.T_profile — Method.

T_profile(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.Tape_L — Method.

Tape_L()

Length of tape per cable [m].

source

Tokamak.Tape_N — Method.

Tape_N()

Number of tapes per 12x12 mm HTS stack.

source

Tokamak.VJ_CS — Function.

VJ_CS(cur_solution=solve_magnet_equations())

Volume of HTS.

source

Tokamak.VM_CS — Function.

VM_CS(cur_solution=solve_magnet_equations())

Volume of Structure.

source

Tokamak.V_TF — Function.

V_TF(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Volume of TF Coil Structure not including HTS.

source

Tokamak.V_inconel — Method.

V_inconel()

Lorem ipsum dolor sit amet.

source

Tokamak.V_w_c — Method.

V_w_c()

Lorem ipsum dolor sit amet.

source

Tokamak.V_w_fl — Method.

V_w_fl()

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_CS — Function.

Vol_CS(cur_solution=solve_magnet_equations())

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_Cu — Method.

Vol_Cu()

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_H2 — Method.

Vol_H2()

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_HTS — Method.

Vol_HTS()

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_HTS_Total — Function.

Vol_HTS_Total(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_PF — Method.

Vol_PF()

Volume of PF Coils (HTS and Structure).

source

Tokamak.Vol_ST_Total — Function.

Vol_ST_Total(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_St — Method.

Vol_St()

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_TF — Method.

Vol_TF()

Lorem ipsum dolor sit amet.

source

Tokamak.Vol_WP — Function.

Vol_WP(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Lorem ipsum dolor sit amet.

source

Tokamak.Vsc_PF — Method.

Vsc_PF()

Lorem ipsum dolor sit amet.

source

Tokamak.Vst_PF — Method.

Vst_PF()

Lorem ipsum dolor sit amet.

source

Tokamak.WP_Area — Function.

WP_Area(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Determine Total Cable Cross Secional Area.

source

Tokamak.WP_d — Method.

WP_d()

Winding pack radial depth [m].

source

Tokamak.WP_w — Method.

WP_w()

Widing pack toroidal width [m].

source

Tokamak.a — Function.

a(cur_R_0=R_0)

Lorem ipsum dolor sit amet.

source

Tokamak.a2CS — Function.

a2CS(cur_solution=solve_magnet_equations())

Outer Radius.

source

Tokamak.a_bar — Method.

a_bar()

Lorem ipsum dolor sit amet.

source

Tokamak.a_hat — Method.

a_hat()

Lorem ipsum dolor sit amet.

source

Tokamak.a_kappa — Method.

a_kappa()

Lorem ipsum dolor sit amet.

source

Tokamak.add_line_item! — Function.

add_line_item!(cur_table::CostTable, name::AbstractString, cur_fields::Dict=Dict())

Lorem ipsum dolor sit amet.

source

Tokamak.analyze_solved_case — Method.

analyze_solved_case(found_data; verbose=true)

Lorem ipsum dolor sit amet.

source

Tokamak.b_theta — Method.

b_theta(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.b_theta_target — Method.

b_theta_target()

Lorem ipsum dolor sit amet.

source

Tokamak.betap — Method.

betap()

Lorem ipsum dolor sit amet.

source

Tokamak.blanket_thickness — Method.

blanket_thickness()

Total thickness of everything between plasma and magnets, ie: first wall, vacuum vessel, flibe blanket, be multiplier, and shielding..

source

Tokamak.calc_cur_eta_cd — Method.

calc_cur_eta_cd(cur_vars...)

Lorem ipsum dolor sit amet.

source

Tokamak.calc_possible_values — Function.

calc_possible_values()

Lorem ipsum dolor sit amet.

source

Tokamak.calculate_total_cost — Method.

calculate_total_cost(cur_item::LineItem)

Lorem ipsum dolor sit amet.

source

Tokamak.confinement_time_from_scaling — Method.

confinement_time_from_scaling()

Lorem ipsum dolor sit amet.

source

Tokamak.construction_distribution — Method.

construction_distribution()

Lorem ipsum dolor sit amet.

source

Tokamak.converge_eta_CD — Method.

converge_eta_CD(cur_equations; verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.cost_schedule — Method.

cost_schedule(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.d_shape_array — Method.

d_shape_array()

Lorem ipsum dolor sit amet.

source

Tokamak.divertor_material_cost — Method.

divertor_material_cost()

Heating due to thermal shielding @ 80K [kW].

source

Tokamak.econ_MW_e — Method.

econ_MW_e()

power output of the plant in MW.

source

Tokamak.econ_V_steel — Method.

econ_V_steel()

Lorem ipsum dolor sit amet.

source

Tokamak.econ_kW_h_per_year — Method.

econ_kW_h_per_year()

Lorem ipsum dolor sit amet.

source

Tokamak.elliptice — Method.

elliptice(cur_val)

Lorem ipsum dolor sit amet.

source

Tokamak.elliptick — Method.

elliptick(cur_val)

Lorem ipsum dolor sit amet.

source

Tokamak.f_B — Method.

f_B()

Lorem ipsum dolor sit amet.

source

Tokamak.f_CD — Method.

f_CD()

Lorem ipsum dolor sit amet.

source

Tokamak.f_RP — Method.

f_RP()

Lorem ipsum dolor sit amet.

source

Tokamak.fill_out_cost_table! — Method.

fill_out_cost_table!(MCT::CostTable, analyzed_solution::Dict)

Lorem ipsum dolor sit amet.

source

Tokamak.geometry_scaling — Method.

geometry_scaling()

Lorem ipsum dolor sit amet.

source

Tokamak.get_new_eta_CD — Method.

get_new_eta_CD(cur_solved_R_0, cur_solved_B_0; verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.greenwald_limit — Method.

greenwald_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.heat_load_limit — Method.

heat_load_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.hts_thickness — Method.

hts_thickness()

Lorem ipsum dolor sit amet.

source

Tokamak.ignition_requirement — Method.

ignition_requirement()

Lorem ipsum dolor sit amet.

source

Tokamak.l_over_2_pi_a — Method.

l_over_2_pi_a()

Lorem ipsum dolor sit amet.

source

Tokamak.lawson_criterion — Method.

lawson_criterion()

Lorem ipsum dolor sit amet.

source

Tokamak.lcoe — Method.

lcoe(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.local_B — Function.

local_B(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Cost_Total — Function.

magnet_Cost_Total(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=nothing)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Fr — Method.

magnet_Fr(ri,rj,zi,zj,Ii,Ij)

Forces on PF and Divertor Coils (Radial Force).

source

Tokamak.magnet_Frc — Method.

magnet_Frc()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Frp — Method.

magnet_Frp(ri,rm,zi,zm,Ii,Im)

Force due to plasma.

source

Tokamak.magnet_Frs — Method.

magnet_Frs(Ii)

Radial Self Force.

source

Tokamak.magnet_Fz — Method.

magnet_Fz(ri,rj,zi,zj,Ii,Ij)

Forces on PF and Divertor Coils (Axial Force).

source

Tokamak.magnet_Fzp — Method.

magnet_Fzp(ri,rm,zi,zm,Ii,Im)

Force due to plasma.

source

Tokamak.magnet_Ic — Method.

magnet_Ic()

Current per coil.

source

Tokamak.magnet_Id_array — Method.

magnet_Id_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Imax — Method.

magnet_Imax()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Ip_array — Method.

magnet_Ip_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_L1 — Method.

magnet_L1()

Straight Section Arc Length.

source

Tokamak.magnet_L2 — Function.

magnet_L2(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Curved Section Arc Length.

source

Tokamak.magnet_Q_15 — Function.

magnet_Q_15(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Total hydrogen cooling @ 15K [kW].

source

Tokamak.magnet_Q_80 — Method.

magnet_Q_80()

Heating due to thermal shielding @ 80K [kW].

source

Tokamak.magnet_Q_total — Function.

magnet_Q_total(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Total cooling equivalent at 4.5 K to use in Green formula [kW].

source

Tokamak.magnet_Turns1 — Method.

magnet_Turns1()

Number of cable turns per coil.

source

Tokamak.magnet_a1PF — Method.

magnet_a1PF()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_a2PF — Method.

magnet_a2PF()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_c_array — Method.

magnet_c_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_centering_force — Function.

magnet_centering_force(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Net Centering Force.

source

Tokamak.magnet_cj — Method.

magnet_cj()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_cm — Method.

magnet_cm()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_cold_mass — Function.

magnet_cold_mass(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Total volume of magnets times density (~8000 kg/m3) in [tons].

source

Tokamak.magnet_etad — Method.

magnet_etad()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_etap — Method.

magnet_etap()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_inner_radius — Method.

magnet_inner_radius()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_inward_centering — Method.

magnet_inward_centering()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_k_array — Method.

magnet_k_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_kd — Method.

magnet_kd()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_kp — Method.

magnet_kp()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_li — Method.

magnet_li()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_material_thickness — Function.

magnet_material_thickness(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_normalized_radius — Method.

magnet_normalized_radius()

Normalized Inner Thickness.

source

Tokamak.magnet_outward_centering — Function.

magnet_outward_centering(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_rd_array — Method.

magnet_rd_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_rhod — Method.

magnet_rhod()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_rhop — Method.

magnet_rhop()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_rp_array — Method.

magnet_rp_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_rpp — Method.

magnet_rpp()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_subs — Function.

magnet_subs(cur_var, cur_R_0, cur_n_bar, cur_I_M, cur_B_0=B_0)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_turns — Method.

magnet_turns()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_z_array — Function.

magnet_z_array(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_zd_array — Method.

magnet_zd_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_zp_array — Method.

magnet_zp_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_zpp — Method.

magnet_zpp()

Lorem ipsum dolor sit amet.

source

Tokamak.n_profile — Method.

n_profile(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.neutron_source_rate — Method.

neutron_source_rate(cur_P_W)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_ce — Method.

omega_ce(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_ci — Method.

omega_ci(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_pe2 — Method.

omega_pe2(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.overnight_cost — Method.

overnight_cost(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.overnight_cost_per_MW — Method.

overnight_cost_per_MW(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.power_balance — Method.

power_balance()

Lorem ipsum dolor sit amet.

source

Tokamak.power_sinks — Method.

power_sinks()

Lorem ipsum dolor sit amet.

source

Tokamak.power_sources — Method.

power_sources()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_beta_limit — Method.

r_b_eq_from_beta_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_heat_loading — Method.

r_b_eq_from_heat_loading()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_ignition — Method.

r_b_eq_from_ignition()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_wall_loading — Method.

r_b_eq_from_wall_loading()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_hard_coded — Method.

r_b_eq_hard_coded()

Lorem ipsum dolor sit amet.

source

Tokamak.reduced_beta — Method.

reduced_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.scan_for_R_0 — Function.

scan_for_R_0(R_0_value, T_list=linspace(5,30,5); rel_tol=1e-3, is_first_call=true, verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.setup_given_equations — Method.

setup_given_equations()

Lorem ipsum dolor sit amet.

source

Tokamak.shield_thickness — Method.

shield_thickness()

Lorem ipsum dolor sit amet.

source

Tokamak.sigma_v — Method.

sigma_v()

Lorem ipsum dolor sit amet.

source

Tokamak.sigma_v_ave — Method.

sigma_v_ave(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.simplified_current — Method.

simplified_current()

Lorem ipsum dolor sit amet.

source

Tokamak.simplified_density — Method.

simplified_density()

Lorem ipsum dolor sit amet.

source

Tokamak.simplified_n_bar_tau_E — Method.

simplified_n_bar_tau_E()

Lorem ipsum dolor sit amet.

source

Tokamak.solenoid_current — Method.

solenoid_current()

Lorem ipsum dolor sit amet.

source

Tokamak.solenoid_length — Method.

solenoid_length()

Lorem ipsum dolor sit amet.

source

Tokamak.solve_equation_set — Method.

solve_equation_set(cur_T, given_equations; verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.solve_magnet_equations — Function.

solve_magnet_equations(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; verbose=false)

Define Inner Radius and Total Thickness.

source

Tokamak.solve_wave_equations — Method.

solve_wave_equations(cur_solved_R_0, cur_solved_B_0; verbose=false)

Function to solve for n_para, rho_J, omega_nor2.

source

Tokamak.solved_B_0_from_R_0 — Method.

solved_B_0_from_R_0()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_B_0_from_beta — Method.

solved_B_0_from_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_B_0_from_heat — Method.

solved_B_0_from_heat()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_B_0_from_wall — Method.

solved_B_0_from_wall()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_B_0 — Method.

solved_R_0_from_B_0(B_0)

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_beta — Method.

solved_R_0_from_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_heat — Method.

solved_R_0_from_heat()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_wall — Method.

solved_R_0_from_wall()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_steady_current — Method.

solved_steady_current()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_steady_density — Method.

solved_steady_density()

Lorem ipsum dolor sit amet.

source

Tokamak.steady_state — Method.

steady_state()

Lorem ipsum dolor sit amet.

source

Tokamak.surface_area — Function.

surface_area(cur_R_0=R_0)

Lorem ipsum dolor sit amet.

source

Tokamak.sweep_T_k — Method.

sweep_T_k(T_list; verbose=true)

Lorem ipsum dolor sit amet.

source

Tokamak.target_angle — Method.

target_angle()

Lorem ipsum dolor sit amet.

source

Tokamak.thermal_power — Method.

thermal_power()

Lorem ipsum dolor sit amet.

source

Tokamak.total_costs_per_year — Method.

total_costs_per_year(cur_table::CostTable)

Discount cost per year.

source

Tokamak.troyon_beta_limit — Method.

troyon_beta_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.update_cost_data_frame! — Method.

update_cost_data_frame!(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.v_e — Method.

v_e(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.volume — Function.

volume(cur_R_0=R_0; offset=0u"m")

Lorem ipsum dolor sit amet.

source

Tokamak.wall_loading_limit — Method.

wall_loading_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.wave_chi — Method.

wave_chi(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.x_aa_of_pi — Method.

x_aa_of_pi()

Lorem ipsum dolor sit amet.

source

Types

Tokamak.CostTable — Type.

CostTable(cur_R_0=R_0, cur_B_0=B_0, cur_T_k=T_k; analyzed_solution=Dict())

Lorem ipsum dolor sit amet.

source

Tokamak.LineItem — Type.

LineItem(name, price_per, quantity, fab_factor)

Lorem ipsum dolor sit amet.

source

Constants

Tokamak.E_Li — Constant.

E_Li

Lorem ipsum dolor sit amet.

source

Tokamak.E_alpha — Constant.

E_alpha

Lorem ipsum dolor sit amet.

source

Tokamak.standard_mu_0 — Constant.

standard_mu_0

Permeability.

source

Macros

Tokamak.@generate_magnet_equation_set — Macro.

@generate_magnet_equation_set

Lorem ipsum dolor sit amet.

source

Tokamak.@generate_wave_equation_set — Macro.

@generate_wave_equation_set

Lorem ipsum dolor sit amet.

source