Code

Code

Functions

Tokamak.B_M — Method.

B_M(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.C_B — Method.

C_B()

Lorem ipsum dolor sit amet.

source

Tokamak.Cryo_Operation — Function.

Cryo_Operation(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Cost of power @ 0.1 /kWhr to power the cryoplant annually.

source

Tokamak.Cryoplant_Cost — Function.

Cryoplant_Cost(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M; cur_solution=solve_magnet_equations())

Cost of cryoplant scaled to various heat loads, scaled to 4.5K and estimated using Green's Formula.

source

Tokamak.Cryostat_Cost — Method.

Cryostat_Cost()

Cost of cryostat scaled to ITER.

source

Tokamak.Cu_Cp — Method.

Cu_Cp(T)

Specific heat of copper as func of temperature.

source

Tokamak.E_F — Method.

E_F()

Lorem ipsum dolor sit amet.

source

Tokamak.E_n — Method.

E_n()

Lorem ipsum dolor sit amet.

source

Tokamak.I_CD — Method.

I_CD()

Lorem ipsum dolor sit amet.

source

Tokamak.K_B — Method.

K_B()

Lorem ipsum dolor sit amet.

source

Tokamak.K_BR — Method.

K_BR()

Lorem ipsum dolor sit amet.

source

Tokamak.K_CD — Method.

K_CD()

Lorem ipsum dolor sit amet.

source

Tokamak.K_CD_hat — Method.

K_CD_hat()

Lorem ipsum dolor sit amet.

source

Tokamak.K_DV — Method.

K_DV()

Lorem ipsum dolor sit amet.

source

Tokamak.K_F — Method.

K_F()

Lorem ipsum dolor sit amet.

source

Tokamak.K_H — Method.

K_H()

Lorem ipsum dolor sit amet.

source

Tokamak.K_I — Method.

K_I()

Lorem ipsum dolor sit amet.

source

Tokamak.K_L — Method.

K_L()

Lorem ipsum dolor sit amet.

source

Tokamak.K_N_G — Method.

K_N_G()

Lorem ipsum dolor sit amet.

source

Tokamak.K_PB — Method.

K_PB()

Lorem ipsum dolor sit amet.

source

Tokamak.K_W — Method.

K_W()

Lorem ipsum dolor sit amet.

source

Tokamak.K_beta — Method.

K_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.K_kappa — Method.

K_kappa()

Lorem ipsum dolor sit amet.

source

Tokamak.K_kernel — Method.

K_kernel()

Lorem ipsum dolor sit amet.

source

Tokamak.K_n — Method.

K_n()

Lorem ipsum dolor sit amet.

source

Tokamak.K_rad — Method.

K_rad()

Lorem ipsum dolor sit amet.

source

Tokamak.MWh_electric_per_year — Method.

MWh_electric_per_year()

Discount electricity per year.

source

Tokamak.P_BR — Method.

P_BR()

Lorem ipsum dolor sit amet.

source

Tokamak.P_F — Method.

P_F()

Lorem ipsum dolor sit amet.

source

Tokamak.P_H — Method.

P_H()

Lorem ipsum dolor sit amet.

source

Tokamak.P_T — Method.

P_T()

Lorem ipsum dolor sit amet.

source

Tokamak.P_alpha — Method.

P_alpha()

Lorem ipsum dolor sit amet.

source

Tokamak.P_kappa — Method.

P_kappa()

Lorem ipsum dolor sit amet.

source

Tokamak.P_n — Method.

P_n()

Lorem ipsum dolor sit amet.

source

Tokamak.Q_E — Method.

Q_E()

Lorem ipsum dolor sit amet.

source

Tokamak.S_W — Method.

S_W()

Lorem ipsum dolor sit amet.

source

Tokamak.T_profile — Method.

T_profile(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.V_inconel — Method.

V_inconel()

Lorem ipsum dolor sit amet.

source

Tokamak.V_w_c — Method.

V_w_c()

Lorem ipsum dolor sit amet.

source

Tokamak.V_w_fl — Method.

V_w_fl()

Lorem ipsum dolor sit amet.

source

Tokamak.a — Function.

a(cur_R_0=R_0)

Lorem ipsum dolor sit amet.

source

Tokamak.a_bar — Method.

a_bar()

Lorem ipsum dolor sit amet.

source

Tokamak.a_hat — Method.

a_hat()

Lorem ipsum dolor sit amet.

source

Tokamak.a_kappa — Function.

a_kappa(cur_R_0=R_0)

Lorem ipsum dolor sit amet.

source

Tokamak.add_line_item! — Function.

add_line_item!(cur_table::CostTable, name::AbstractString, cur_fields::Dict=Dict())

Lorem ipsum dolor sit amet.

source

Tokamak.analyze_solved_case — Method.

analyze_solved_case(found_data; verbose=true)

Lorem ipsum dolor sit amet.

source

Tokamak.b_theta — Method.

b_theta(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.b_theta_target — Method.

b_theta_target()

Lorem ipsum dolor sit amet.

source

Tokamak.blanket_thickness — Function.

blanket_thickness(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Total thickness of everything between plasma and magnets, ie: first wall, vacuum vessel, flibe blanket, be multiplier, and shielding..

source

Tokamak.calc_cur_eta_cd — Method.

calc_cur_eta_cd(cur_vars...)

Lorem ipsum dolor sit amet.

source

Tokamak.calc_possible_values — Function.

calc_possible_values()

Lorem ipsum dolor sit amet.

source

Tokamak.calculate_total_cost — Method.

calculate_total_cost(cur_item::LineItem)

Lorem ipsum dolor sit amet.

source

Tokamak.confinement_time_from_scaling — Method.

confinement_time_from_scaling()

Lorem ipsum dolor sit amet.

source

Tokamak.construction_distribution — Method.

construction_distribution()

Lorem ipsum dolor sit amet.

source

Tokamak.converge_eta_CD — Function.

converge_eta_CD(cur_equations, T_guess=12.5; verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.cost_schedule — Method.

cost_schedule(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.d_shape_array — Method.

d_shape_array()

Lorem ipsum dolor sit amet.

source

Tokamak.dememoize — Method.

dememoize()

Lorem ipsum dolor sit amet.

source

Tokamak.divertor_material_cost — Method.

divertor_material_cost()

Heating due to thermal shielding @ 80K [kW].

source

Tokamak.econ_MW_e — Method.

econ_MW_e()

power output of the plant in MW.

source

Tokamak.econ_V_steel — Method.

econ_V_steel()

Lorem ipsum dolor sit amet.

source

Tokamak.econ_kW_h_per_year — Method.

econ_kW_h_per_year()

Lorem ipsum dolor sit amet.

source

Tokamak.elliptice — Method.

elliptice(cur_val)

Lorem ipsum dolor sit amet.

source

Tokamak.elliptick — Method.

elliptick(cur_val)

Lorem ipsum dolor sit amet.

source

Tokamak.f_B — Method.

f_B()

Lorem ipsum dolor sit amet.

source

Tokamak.f_CD — Method.

f_CD()

Lorem ipsum dolor sit amet.

source

Tokamak.f_RP — Method.

f_RP()

Lorem ipsum dolor sit amet.

source

Tokamak.fill_out_cost_table! — Method.

fill_out_cost_table!(MCT::CostTable, analyzed_solution::Dict)

Lorem ipsum dolor sit amet.

source

Tokamak.geometry_scaling — Method.

geometry_scaling()

Lorem ipsum dolor sit amet.

source

Tokamak.get_new_eta_CD — Method.

get_new_eta_CD(cur_solved_R_0, cur_solved_T_k; verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.greenwald_limit — Method.

greenwald_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.heat_load_limit — Method.

heat_load_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.ignition_requirement — Method.

ignition_requirement()

Lorem ipsum dolor sit amet.

source

Tokamak.l_over_2_pi_a — Method.

l_over_2_pi_a()

Lorem ipsum dolor sit amet.

source

Tokamak.lawson_criterion — Method.

lawson_criterion()

Lorem ipsum dolor sit amet.

source

Tokamak.lcoe — Method.

lcoe(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.lin_sigma_v — Method.

lin_sigma_v()

Lorem ipsum dolor sit amet.

source

Tokamak.lin_sigma_v_ave — Method.

lin_sigma_v_ave(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.local_B — Function.

local_B(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.log_sigma_v — Method.

log_sigma_v()

Lorem ipsum dolor sit amet.

source

Tokamak.log_sigma_v_ave — Method.

log_sigma_v_ave(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Id_array — Method.

magnet_Id_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_Ip_array — Method.

magnet_Ip_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_zd_array — Method.

magnet_zd_array()

Lorem ipsum dolor sit amet.

source

Tokamak.magnet_zp_array — Method.

magnet_zp_array()

Lorem ipsum dolor sit amet.

source

Tokamak.n_para — Function.

n_para(rho_J, cur_wave_chi=nothing)

Lorem ipsum dolor sit amet.

source

Tokamak.n_profile — Method.

n_profile(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.neutron_source_rate — Method.

neutron_source_rate(cur_P_W)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_ce — Method.

omega_ce(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_ci — Method.

omega_ci(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_nor2 — Function.

omega_nor2(rho_J, cur_wave_chi=nothing)

Lorem ipsum dolor sit amet.

source

Tokamak.omega_pe2 — Method.

omega_pe2(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.overnight_cost — Method.

overnight_cost(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.overnight_cost_per_MW — Method.

overnight_cost_per_MW(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.power_balance — Method.

power_balance()

Lorem ipsum dolor sit amet.

source

Tokamak.power_sinks — Method.

power_sinks()

Lorem ipsum dolor sit amet.

source

Tokamak.power_sources — Method.

power_sources()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_beta_limit — Method.

r_b_eq_from_beta_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_heat_loading — Method.

r_b_eq_from_heat_loading()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_ignition — Method.

r_b_eq_from_ignition()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_from_wall_loading — Method.

r_b_eq_from_wall_loading()

Lorem ipsum dolor sit amet.

source

Tokamak.r_b_eq_hard_coded — Method.

r_b_eq_hard_coded()

Lorem ipsum dolor sit amet.

source

Tokamak.reduced_beta — Method.

reduced_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.scan_for_R_0 — Function.

scan_for_R_0(R_0_value, T_list=linspace(5,30,5); rel_tol=1e-3, is_first_call=true, verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.setup_given_equations — Method.

setup_given_equations()

Lorem ipsum dolor sit amet.

source

Tokamak.shield_thickness — Function.

shield_thickness(cur_R_0=R_0, cur_n_bar=n_bar, cur_I_M=I_M)

Lorem ipsum dolor sit amet.

source

Tokamak.sigma_v — Method.

sigma_v()

Lorem ipsum dolor sit amet.

source

Tokamak.sigma_v_ave — Method.

sigma_v_ave(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.sigma_v_k_array — Method.

sigma_v_k_array()

Lorem ipsum dolor sit amet.

source

Tokamak.simplified_current — Method.

simplified_current()

Lorem ipsum dolor sit amet.

source

Tokamak.simplified_density — Method.

simplified_density()

Lorem ipsum dolor sit amet.

source

Tokamak.simplified_n_bar_tau_E — Method.

simplified_n_bar_tau_E()

Lorem ipsum dolor sit amet.

source

Tokamak.solve_equation_set — Method.

solve_equation_set(cur_B, given_equations; verbose=false)

Lorem ipsum dolor sit amet.

source

Tokamak.solve_wave_equations — Method.

solve_wave_equations(cur_solved_R_0, cur_solved_T_k; verbose=false)

Function to solve for n_para, rho_J, omega.

source

Tokamak.solved_B_0_from_R_0 — Method.

solved_B_0_from_R_0()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_B_0_from_beta — Method.

solved_B_0_from_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_B_0_from_heat — Method.

solved_B_0_from_heat()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_B_0_from_wall — Method.

solved_B_0_from_wall()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_B_0 — Method.

solved_R_0_from_B_0(B_0)

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_beta — Method.

solved_R_0_from_beta()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_heat — Method.

solved_R_0_from_heat()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_power_balance — Function.

solved_R_0_from_power_balance(cur_T_k, cur_B_0)

Lorem ipsum dolor sit amet.

source

Tokamak.solved_R_0_from_wall — Method.

solved_R_0_from_wall()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_T_k_from_beta — Function.

solved_T_k_from_beta(T_guess, cur_B_0=B_0)

Lorem ipsum dolor sit amet.

source

Tokamak.solved_steady_current — Method.

solved_steady_current()

Lorem ipsum dolor sit amet.

source

Tokamak.solved_steady_density — Method.

solved_steady_density()

Lorem ipsum dolor sit amet.

source

Tokamak.steady_state — Method.

steady_state()

Lorem ipsum dolor sit amet.

source

Tokamak.surface_area — Function.

surface_area(cur_R_0=R_0)

Lorem ipsum dolor sit amet.

source

Tokamak.sweep_B_0 — Method.

sweep_B_0(B_list; verbose=true)

Lorem ipsum dolor sit amet.

source

Tokamak.target_angle — Method.

target_angle()

Lorem ipsum dolor sit amet.

source

Tokamak.thermal_power — Method.

thermal_power()

Lorem ipsum dolor sit amet.

source

Tokamak.total_costs_per_year — Method.

total_costs_per_year(cur_table::CostTable)

Discount cost per year.

source

Tokamak.troyon_beta_limit — Method.

troyon_beta_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.update_cost_data_frame! — Method.

update_cost_data_frame!(cur_table::CostTable)

Lorem ipsum dolor sit amet.

source

Tokamak.v_e — Method.

v_e(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.volume — Function.

volume(cur_R_0=R_0; offset=0u"m")

Lorem ipsum dolor sit amet.

source

Tokamak.wall_loading_limit — Method.

wall_loading_limit()

Lorem ipsum dolor sit amet.

source

Tokamak.wave_chi — Method.

wave_chi(rho)

Lorem ipsum dolor sit amet.

source

Tokamak.x_aa_of_pi — Method.

x_aa_of_pi()

Lorem ipsum dolor sit amet.

source

Types

Tokamak.CostTable — Type.

CostTable(cur_R_0=R_0, cur_B_0=B_0, cur_T_k=T_k; analyzed_solution=Dict(), magnet_solution=nothing)

Lorem ipsum dolor sit amet.

source

Tokamak.LineItem — Type.

LineItem(name, price_per, quantity, fab_factor)

Lorem ipsum dolor sit amet.

source

Constants

Tokamak.E_Li — Constant.

E_Li

Lorem ipsum dolor sit amet.

source

Tokamak.E_alpha — Constant.

E_alpha

Lorem ipsum dolor sit amet.

source

Tokamak.standard_mu_0 — Constant.

standard_mu_0

Permeability.

source

Macros

Tokamak.@generate_magnet_equation_set — Macro.

@generate_magnet_equation_set

Lorem ipsum dolor sit amet.

source

Tokamak.@generate_wave_equation_set — Macro.

@generate_wave_equation_set

Lorem ipsum dolor sit amet.

source